Rename "kickOn" method to "kickWhen"
For some reason, I wasn't happy with naming the method "kickOn", whereas
this one sounds more appropriate somehow.

Also fixed a lingering instance of an earlier name for the method, which
was still mentioned in the preceding comment block ("resetOn"). I should
have picked up on that in 52ab564, but it escaped my notice.