Fix inability to disable automatic nagging
A dumb oversight on my part made it impossible to disable timed triggers
by passing a falsy value to "showAfter". I meant for a default of 4000ms
to be used if the option wasn't specified in the constructor, but didn't
pay too much attention to what the hell I was writing.

Resolves #2.