Add options to configure cookie attributes
When a user closes a Nag, a cookie is created to record the dismissal so
it won't bug them a second time. However, the cookie was set without any
expiration, meaning it was reset once the user's browsing session ended.
Since this would inevitably drive users insane after repeated visits, an
option to set the cookie's lifespan was needed.

As I was extending the list of supported arguments to include one cookie
property, I realised I may as well include the others too. The following
options now set their corresponding RFC 6265 cookie attribute:

    * cookieDomain:    domain
    * cookieExpires:   expires
    * cookiePath:      path
    * cookieSecure:    secure