Reverse logic of kickWhen method
It makes more sense in a lexical context for the element selectors to be
keyed to event types, rather than the other way around. For instance, we
now declare how Nags are dismissed in an almost imperative tone:

    kickWhen: {
        click: "#close-button",
        submit: "form"
    }

Which could be read by a developer as "Kick when click close button" and
"Kick when submit form", respectively. More importantly, keying elements
to event types obviates the need to quote each property's name in object
literals. For instance, compare the above example with this:

    kickWhen: {
        "#close-button": "click",
        "form": "submit"
    }

Furthermore, reversing the logic this way now enables developers to pass
an array of selectors if they so desire. Since arrays already join their
values with commas when stringified, it makes no difference which of the
two data types they use:

    kickWhen: { click:  ".close-btn,   .cancel"  }
    kickWhen: { click: [".close-btn", ".cancel"] }

Fixes #3.