Refactor how kick listeners are registered
The recently-added "softlyKickWhen" method's been removed in favour of a
more streamlined (and logical) approach. Instead of setting the handlers
with methods, the options hash passed to the constructor now accepts two
additional parameters:

    new Nag(element, {
        kickWhen:       { click: "#close-btn" },
        softlyKickWhen: { submit: "form" }
    });

Which makes a lot more sense than forcing one to do this instead:

    new Nag(element)
            .kickWhen(       { click: "#close-btn" })
            .softlyKickWhen( { submit: "form"      });

Note that the "kickWhen" method is still available for convenience, with
the "soft" parameter added as an optional second argument to replace the
now-deleted "softlyKickWhen".

Really not sure what I was thinking when I originally wrote this.