Add lazy initialisation alternatives
To simplify basic usage, developers can now create new Nags by passing a
string instead of an element reference if they desire. For example, both
of the following statements achieve the same thing:

    new Nag(document.querySelector("#popup"));
    new Nag("#popup");

Secondly, the second argument can be replaced with a string value, which
is interpreted as the option hash's "cookieName" property instead:

    new Nag("#popup", {cookieName: "seen-popup"});
    new Nag("#popup", "seen-popup");

Finally, there's now a third parameter called "lazy" which intelligently
adds kick-handlers to any elements with obvious-looking classes:

    new Nag("#popup", "seen-popup", true);

The above line is functionally identical to this:

    new Nag("#popup", {
        cookieName:      "seen-popup",
        kickWhen:       { click: ".close, .cancel, .close-btn"},
        kickSoftlyWhen: { submit: "form" }
    });

If only one parameter was passed to the constructor, the "lazy" option's
turned on automatically. This is done because it can be safely assumed a
developer not bothering to even include the Nag's cookieName is probably
too lazy to be specific in which elements dismiss the nag when clicked.

These changes are all inspired by many precedent scenarios where minimal
configuration is neither desired nor necessary. Remember these alternate
initialiser methods aren't mandatory; they merely offer tired, irritated
coders a quicker way to achieve what's otherwise a really basic effect.

Closes #6.